+++ src/index.js
@@ -25,6 +25,7 @@ import * as pressOn from './press-on';
 import typeText from './type-text';
 import isTextSelected from './is-text-selected';
 import deleteSelection from './delete-selection';
+import isDocumentActive from './is-document-active';
 
 
 export {
@@ -47,5 +48,6 @@ export {
     pressOn,
     typeText,
     isTextSelected,
-    deleteSelection
+    deleteSelection,
+    isDocumentActive
 };