+++ types/index.d.ts
@@ -97,6 +97,8 @@ export function typeText(text: string): void;
 
 export function isTextSelected(): boolean;
 
+export function isDocumentActive(): boolean;
+
 export function deleteSelection(): boolean;
 
 //#endregion